+++ skoolkit/skool2sft.py
@@ -23,12 +23,12 @@ from .skoolsft import SftWriter
 
 def run(skoolfile, options):
     writer = SftWriter(skoolfile, options.write_hex, options.preserve_base)
-    writer.write()
+    writer.write(options.start, options.end)
 
 def main(args):
     parser = argparse.ArgumentParser(
         usage='skool2sft.py [options] FILE',
-        description="Convert a skool file into a skool file template, written to standard output. "
+        description="Convert a skool file into a skool file template and write it to standard output. "
                     "FILE may be a regular file, or '-' for standard input.",
         add_help=False
     )
@@ -36,10 +36,14 @@ def main(args):
     group = parser.add_argument_group('Options')
     group.add_argument('-b', '--preserve-base', action='store_true', dest='preserve_base',
                        help="Preserve the base of decimal and hexadecimal values in instruction operands and DEFB/DEFM/DEFS/DEFW statements")
+    group.add_argument('-E', '--end', dest='end', metavar='ADDR', type=int, default=65536,
+                       help="Stop converting at this address")
     group.add_argument('-h', '--hex', action='store_const', dest='write_hex', const=1, default=0,
                        help='Write addresses in upper case hexadecimal format')
     group.add_argument('-l', '--hex-lower', action='store_const', dest='write_hex', const=-1, default=0,
                        help='Write addresses in lower case hexadecimal format')
+    group.add_argument('-S', '--start', dest='start', metavar='ADDR', type=int, default=0,
+                       help="Start converting at this address")
     group.add_argument('-V', '--version', action='version', version='SkoolKit {}'.format(VERSION),
                        help='Show SkoolKit version number and exit')
     namespace, unknown_args = parser.parse_known_args(args)