+++ tests/test_skool2sft.py
@@ -15,8 +15,10 @@ class MockSftWriter:
         self.write_called = False
         mock_sft_writer = self
 
-    def write(self):
+    def write(self, min_address, max_address):
         self.write_called = True
+        self.min_address = min_address
+        self.max_address = max_address
 
 class Skool2SftTest(SkoolKitTestCase):
     def test_no_arguments(self):
@@ -39,6 +41,8 @@ class Skool2SftTest(SkoolKitTestCase):
         self.assertEqual(write_hex, 0)
         self.assertFalse(preserve_base)
         self.assertTrue(mock_sft_writer.write_called)
+        self.assertEqual(mock_sft_writer.min_address, 0)
+        self.assertEqual(mock_sft_writer.max_address, 65536)
 
     def test_option_V(self):
         for option in ('-V', '--version'):
@@ -77,6 +81,31 @@ class Skool2SftTest(SkoolKitTestCase):
             self.assertEqual(write_hex, 0)
             self.assertTrue(preserve_base)
             self.assertTrue(mock_sft_writer.write_called)
+
+    @patch.object(skool2sft, 'SftWriter', MockSftWriter)
+    def test_option_E(self):
+        for option, end in (('-E', 30000), ('--end', 40000)):
+            skool2sft.main((option, str(end), 'test.skool'))
+            self.assertTrue(mock_sft_writer.write_called)
+            self.assertEqual(mock_sft_writer.min_address, 0)
+            self.assertEqual(mock_sft_writer.max_address, end)
+
+    @patch.object(skool2sft, 'SftWriter', MockSftWriter)
+    def test_option_S(self):
+        for option, start in (('-S', 40000), ('--start', 50000)):
+            skool2sft.main((option, str(start), 'test.skool'))
+            self.assertTrue(mock_sft_writer.write_called)
+            self.assertEqual(mock_sft_writer.min_address, start)
+            self.assertEqual(mock_sft_writer.max_address, 65536)
+
+    @patch.object(skool2sft, 'SftWriter', MockSftWriter)
+    def test_options_S_and_E(self):
+        start = 32768
+        end = 49152
+        skool2sft.main(('-S', str(start), '-E', str(end), 'test.skool'))
+        self.assertTrue(mock_sft_writer.write_called)
+        self.assertEqual(mock_sft_writer.min_address, start)
+        self.assertEqual(mock_sft_writer.max_address, end)
 
     def test_run(self):
         skool = '; Routine\nc32768 RET'